Stop using delete_on_termination for BFV instances
When using delete_on_termination and the booting of the instance fails
on the first attempt, the second attempt will fail with Heat.  The
reason is that with delete_on_termination set to True, Nova will delete
the volume when Heat deletes the ERROR'd instance and it will then
result in the follow-up boot to fail with an error along the line of
unable to find volume, which masks the real failure from the user (which
could potentialy be aquota issue).

With this patch, we no longer set this and instead use the default of
false.  This will not mean we will leak volumes because when we delete
the stack, Heat will do all the right things and delete them in order,
making sure the volume disappears eventually.

Change-Id: I362cea7bf57825035d13d234d0181a2b1fca5743
(cherry picked from commit 2c63aca8c6afbcc2e32a09cc7f3ad2f18265698f)