WIP DNM: Add image discovery capability to capi driver
Currently the new capi driver is not able to discover images by the
name provided in the template image_id, rather it must have the full
UUID of the image. This means it's tricky to configure zuul and other
CI to point to an image when the UUID of the image is not known prior to
running the workflow. This patch adds this functionality.

story: 2009780
Change-Id: I05086b1769cc1dac07d3e092468b0c4d08a04101